backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation & Models
pydantic==2.5.0
pydantic-ai==0.2.19
pydantic-settings==2.1.0

# HTTP Client & Async
httpx==0.25.2
aiofiles==23.2.0

# Environment & Configuration
python-dotenv==1.0.0

# Image Processing & OCR
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1

# AI Services
openai==1.3.0
anthropic==0.7.0
mistralai==1.0.0  # For Mistral OCR

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Observability
logfire==0.51.0

# Database (if needed)
sqlalchemy==2.0.23
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Production
gunicorn==21.2.0
redis==5.0.1  # For caching

# Security Scanning
bandit==1.7.5
safety==2.3.5

# Load Testing
locust==2.17.0